requirements.txt
affine==2.3.0
aiohttp==3.7.4.post0
amqp==1.4.9
anyjson==0.3.3
africastalking==1.2.4
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
billiard==3.3.0.23
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cycler==0.10.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.2.3
django-celery==3.3.1
django-environ==0.4.5
django-filter==2.4.0
django-geojson==3.2.0
django-heroku==0.3.1
django-leaflet==0.28.0
django-redis==4.12.1
django-tables2==2.4.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
esda==2.3.6
folium==0.12.1
geographiclib==1.50
geoip2==4.2.0
geopy
giddy==2.3.3
idna==2.10
inequality==1.0.0
Jinja2==3.0.1
joblib==1.0.1
jsonfield==3.1.0
kiwisolver==1.3.1
kombu==3.0.37
libpysal==4.4.0
llvmlite==0.36.0
mapclassify==2.4.2
MarkupSafe==2.0.1
matplotlib==3.4.2
maxminddb==2.0.3
mgwr==2.1.2
mpmath==1.2.1
multidict==5.1.0
munch==2.5.0
networkx==2.5.1
numba==0.53.1
numpy==1.20.3
opencv-contrib-python==4.5.2.52
pandas==1.2.4
patsy==0.5.1
Pillow==8.2.0
pointpats==2.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pygeos==0.10
pyparsing==2.4.7
python-dateutil==2.8.0
pytz==2021.1
quantecon==0.5.0
rasterstats==0.15.0
redis==3.5.3
reportlab==3.5.67
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
segregation==1.5.0
Shapely==1.7.1
simplejson==3.17.2
six==1.16.0
snuggs==1.4.7
soupsieve==2.2.1
spaghetti==1.5.6
spglm==1.0.8
spint==1.0.7
splot==1.1.3
spreg==1.2.2
spvcm==0.3.0
sqlparse==0.4.1
statsmodels==0.12.2
sympy==1.8
threadpoolctl==2.1.0
tobler==0.7.0
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
whitenoise==5.2.0
yarl==1.6.3
